% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{birch.df}
\alias{birch.df}
\alias{birch_dw.df}
\title{Birch seedlings' size}
\format{
A \code{data.frame} object with 350 rows and 8 variables.

A \code{data.frame} object with 700 rows and 5 variables.
}
\usage{
birch.df

birch_dw.df
}
\description{
A dataset containing the measurements on 350 birch seedlings.
}
\details{
The data are for seedlings grown in trays with cells or containers
of two different volumes. For each of these types of trays, all cells, 1/2 of
the cells or 1/4 of the cells contained seedlings. Root-collar diameter (mm),
height (cm), dry mass (mg) of stems and roots. Measurements done at the end
of the first growing season, after leaf fall.
}
\examples{
colnames(birch.df)
head(birch.df)

colnames(birch_dw.df)
head(birch_dw.df)

}
\references{
Aphalo, P. J. and Rikala, R. (2003) Field performance of
silver-birch planting-stock grown at different spacing and in containers of
different volume. \emph{New Forests}, 25:93-108. \doi{10.1023/A:1022618810937}.
}
\seealso{
Other Plant growth and morphology data:
\code{\link{ivy.df}}
}
\concept{Plant growth and morphology data}
\keyword{datasets}
