% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_default_colour}
\alias{check_default_colour}
\title{Check default.colour argument}
\usage{
check_default_colour(default.colour, theme = NULL)
}
\arguments{
\item{default.colour}{The value to be checked}

\item{theme}{The theme from where to fetch a replacement value when
\code{default.colour} is \code{NULL}.}
}
\value{
A colour definition or NA
}
\description{
if possible use ink from theme as default colour
}
\details{
The replacement value is fetched from the theme only if
  'ggplot2' >= 4.0.0 and otherwise \code{"black"} is used. \code{NA}
  values are passed through as they can be used to disable rendering
  of grob elements.
}
\keyword{internal}
