% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point-linked.r
\name{geom_point_linked}
\alias{geom_point_linked}
\title{Points linked by a segment or arrow}
\usage{
geom_point_linked(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrow = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. If specified and
\code{inherit.aes = TRUE} (the default), is combined with the default
mapping at the top level of the plot. You only need to supply
\code{mapping} if there isn't a mapping defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on the data for this layer,
as a string.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. There
  are three types of arguments you can use here:

  \itemize{ \item Aesthetics: to set an aesthetic to a fixed value, like
  \code{colour = "red"} or \code{size = 3}. \item Other arguments to the
  layer, for example you override the default \code{stat} associated with the
  layer. \item Other arguments passed on to the stat. }}

\item{arrow}{specification for arrow heads, as created by
\code{\link[grid]{arrow}}}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a
warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
A plot layer instance.
}
\description{
The "point_linked" geom is used to display displacement of points in
scatterplots.
}
\examples{

ggplot(mpg[1:20, ],
       aes(cyl,
           hwy,
           label = drv)) +
  geom_point_linked(position = position_nudge_keep(x = 0.2),
                    color = "red", segment.colour = "brown") +
  geom_point()

ggplot(mpg[1:50, ],
       aes(cyl,
           hwy,
           label = drv)) +
  geom_point_linked(position = position_jitter_and_nudge(width = 0.66, height = 2,
                                                         seed = 456,
                                                  nudge.from = "jittered",
                                                  returned.origin = "original"),
                    color = "red", arrow = grid::arrow(length = grid::unit(0.4, "lines"))) +
  geom_point()

}
\seealso{
[ggplot2::geom_point], [nudge_keep] and [nudge_to]
}
