% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge-dodge.R,
%   R/position-nudge-dodge2.R
\name{position_dodge_and_nudge}
\alias{position_dodge_and_nudge}
\alias{position_dodge2_and_nudge}
\title{Combined positions dodge and nudge}
\usage{
position_dodge_and_nudge(
  width = 1,
  preserve = c("total", "single"),
  x = 0,
  y = 0,
  direction = "split"
)

position_dodge2_and_nudge(
  width = 1,
  preserve = c("total", "single"),
  padding = 0.1,
  reverse = FALSE,
  x = 0,
  y = 0,
  direction = "none"
)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the total width of all elements at a
position, or the width of a single element?.}

\item{x, y}{Amount of vertical and horizontal distance to move. A numeric
vector of length 1, or of the same length as rows there are in `data`,}

\item{direction}{One of "none" or "split". A value of "none" replicates the
behavior of [ggplot2::position_nudge]. At the moment "split" changes the
sign of the nudge at zero, which is suiatble for column plots with negative
slices.}

\item{padding}{Padding between elements at the same position. Elements are
shrunk by this proportion to allow space between them. Defaults to 0.1.}

\item{reverse}{If TRUE, will reverse the default stacking order. This is
useful if you're rotating both the plot and legend.}
}
\value{
A \code{"Position"} object.
}
\description{
`position_dodge_and_nudge()` is useful when labelling plots such as grouped
bars, columns, etc. In contrast to [ggplot2::position_nudge],
`position_dodge_and_nudge()` returns in `data` both the original coordinates
and the nudged coordinates.
}
\details{
This position function is backwards compatible with [ggplot2::position_nudge]
but extends it by adding support for stacking and for the repulsive
geometries from package 'ggrepel'.

The wrapper `position_nudge_keep()` with exactly the same signature and
behaviour as [ggplot2::position_nudge] provides an easier to remember name
when the desire is only to have access to both the original and nudged
coordinates.
}
\examples{
df <- data.frame(x1 = c(1, 2, 1, 3, -1),
                 x2 = c("a", "a", "b", "b", "b"),
                 grp = c("some long name", "other name", "some name",
                         "another name", "some long name"))

# Add labels to a horizontal column plot (stacked by default)
ggplot(data = df, aes(x1, x2, group = grp)) +
  geom_col(aes(fill = grp), width = 1,
           position = position_dodge()) +
  geom_vline(xintercept = 0) +
  geom_text(
    aes(label = grp),
    position = position_dodge_and_nudge(x = 0.09),
    angle = 90) +
  theme(legend.position = "none")

ggplot(data = df, aes(x2, x1, group = grp)) +
  geom_col(aes(fill = grp), width = 0.75,
           position = position_dodge(width = 0.75)) +
  geom_vline(xintercept = 0) +
  geom_text(
    aes(label = grp),
    position = position_dodge_and_nudge(y = 0.1, width = 0.75)) +
  theme(legend.position = "none")

}
\seealso{
[ggplot2::position_nudge()], [ggrepel::position_nudge_repel()].

Other position adjustments: 
\code{\link{position_nudge_center}()},
\code{\link{position_nudge_line}()},
\code{\link{position_nudge_to}()},
\code{\link{position_stack_and_nudge}()}
}
\author{
Michał Krassowski \url{http://linkedin.com/in/michal-krassowski/} and
  Pedro J. Aphalo.
}
\concept{position adjustments}
