% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge-stack.R
\name{position_stack_and_nudge}
\alias{position_stack_and_nudge}
\title{Combined positions nudge and stack}
\usage{
position_stack_and_nudge(x = 0, y = 0, vjust = 1, reverse = FALSE)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move. A numeric
vector of length 1, or of the same length as rows there are in `data`,}

\item{vjust}{Vertical adjustment for geoms that have a position (like points
or lines), not a dimension (like bars or areas). Set to 0 to align with the
bottom, 0.5 for the middle, and 1 (the default) for the top.}

\item{reverse}{If TRUE, will reverse the default stacking order. This is
useful if you're rotating both the plot and legend.}
}
\value{
A \code{"Position"} object.
}
\description{
`position_stack_and_nudge()` is useful when labelling plots such as stacked
bars, stacked columns, stacked lines, etc. In contrast to
[ggplot2::position_nudge], `position_stack_and_nudge()` returns in `data`
both the original coordinates and the nudged coordinates.
}
\details{
This position function is backwards compatible with [ggplot2::position_nudge]
but extends it by adding support for stacking and for the repulsive geometries
from package 'ggrepel'.

The wrapper `position_nudge_keep()` with exactly the same signature and
behaviour as [ggplot2::position_nudge] provides an easier to remember name
when the desire is only to have access to both the original and nudged
coordinates.
}
\examples{
df <- data.frame(y = c("a", "a", "b", "b", "b"),
                 x = c(1, 2, 1, 3, -1),
                 grp = c("some long name", "other name", "some name",
                         "another name", "some long name"))

ggplot(data = df, aes(x, y, group = grp)) +
  geom_col(aes(fill = grp), width=0.5) +
  geom_vline(xintercept = 0) +
  geom_text(
    aes(label = grp),
    position = position_stack_and_nudge(vjust = 0.5, y = 0.3)) +
  theme(legend.position = "none")

ggplot(data = df, aes(y, x, group = grp)) +
  geom_col(aes(fill = grp), width=0.5) +
  geom_vline(xintercept = 0) +
  geom_text(
    aes(label = grp),
    position = position_stack_and_nudge(vjust = 0.5, x = -0.3),
    angle = 90) +
  theme(legend.position = "none")

ggplot(data = df, aes(y, x, group = grp)) +
  geom_col(aes(fill = grp), width=0.5) +
  geom_vline(xintercept = 0) +
  geom_text(
    aes(label = grp),
    position = position_stack_and_nudge(vjust = 1, y = -0.2)) +
  theme(legend.position = "none")

ggplot(data = df, aes(x, y, group = grp)) +
  geom_col(aes(fill = grp), width=0.5) +
  geom_vline(xintercept = 0) +
  geom_text_linked(
    aes(label = grp),
    position = position_stack_and_nudge(vjust = 0.5, y = 0.4),
    vjust = "bottom") +
  theme(legend.position = "none")

}
\seealso{
[ggplot2::position_nudge()], [ggrepel::position_nudge_repel()].

Other position adjustments: 
\code{\link{position_nudge_center}()},
\code{\link{position_nudge_line}()},
\code{\link{position_nudge_to}()}
}
\author{
Michał Krassowski \url{http://linkedin.com/in/michal-krassowski/}
  (revised by Pedro J. Aphalo)
}
\concept{position adjustments}
