% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-grob.R, R/geom-label-npc.r,
%   R/geom-margin-arrow.r, R/geom-margin-grob.r, R/geom-margin-point.r,
%   R/geom-plot.R, R/geom-quadrant-lines.R, R/geom-table.R,
%   R/geom-text-linked.r, R/geom-text-npc.r, R/position-nudge-center.R,
%   R/position-nudge-line.R, R/position-nudge-to.R, R/stat-apply.R,
%   R/stat-dens1d-filter.r, R/stat-dens1d-labels.r, R/stat-dens2d-filter.r,
%   R/stat-dens2d-labels.r, R/stat-fmt-tb.R, R/stat-quadrant-counts.R
\docType{data}
\name{grob_draw_panel_fun}
\alias{grob_draw_panel_fun}
\alias{GeomGrob}
\alias{grobnpc_draw_panel_fun}
\alias{GeomGrobNpc}
\alias{GeomLabelNpc}
\alias{GeomXMarginArrow}
\alias{GeomYMarginArrow}
\alias{GeomXMarginGrob}
\alias{GeomYMarginGrob}
\alias{GeomXMarginPoint}
\alias{GeomYMarginPoint}
\alias{gplot_draw_panel_fun}
\alias{GeomPlot}
\alias{gplotnpc_draw_panel_fun}
\alias{GeomPlotNpc}
\alias{GeomQuadrantLines}
\alias{GeomVHLines}
\alias{gtb_draw_panel_fun}
\alias{GeomTable}
\alias{gtbnpc_draw_panel_fun}
\alias{GeomTableNpc}
\alias{GeomLinkedText}
\alias{GeomTextNpc}
\alias{PositionNudgeCenter}
\alias{quant_compute_panel}
\alias{PositionNudgeLine}
\alias{PositionNudgeTo}
\alias{stat_apply_fun}
\alias{Stats}
\alias{StatApplyGroup}
\alias{StatApplyPanel}
\alias{StatDens1dFilter}
\alias{StatDens1dFilterG}
\alias{StatDens1dLabels}
\alias{StatDens2dFilter}
\alias{StatDens2dFilterG}
\alias{StatDens2dLabels}
\alias{fmt_tb_compute_group_fun}
\alias{StatFmtTb}
\alias{compute_counts_fun}
\alias{StatQuadrantCounts}
\title{\code{Stat*} Objects}
\description{
All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.

All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.
}
\details{
Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object, descended
from the top-level \code{Stat}, and each implements various methods and
fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:

Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object, descended
from the top-level \code{Stat}, and each implements various methods and
fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:
}
\seealso{
\code{\link[ggplot2]{ggplot2-ggproto}}

\code{\link[ggplot2]{ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
