% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{ivy.df}
\alias{ivy.df}
\title{Ivy photosynthesis light response}
\format{
A \code{data.frame} object with 36 rows and 6 variables.
}
\usage{
ivy.df
}
\description{
A dataset containing photosynthesis measurements on four ivy plants.
}
\details{
For each plant a light response curve of photosynthesis was measured
using a custom-built system and software that allowed controlling the
concentrations of water vapour and carbon dioxide at the surface of the
leaves, i.e., inside the air boundary layer.
}
\examples{
colnames(ivy.df)
head(ivy.df)

}
\references{
Aphalo, P. J. (1991) Interactions in Stomatal Function. PhD
thesis, University of Edinburgh. \url{http://hdl.handle.net/1842/14758}.
}
\seealso{
Other Plant growth and morphology data: 
\code{\link{birch.df}}
}
\concept{Plant growth and morphology data}
\keyword{datasets}
