% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{volcano_example.df}
\alias{volcano_example.df}
\title{Gene expression data}
\format{
A \code{data.frame} object with 1218 rows and 5 variables
}
\usage{
volcano_example.df
}
\description{
A dataset containing reshaped and simplified output from an
  analysis of data from RNAseq done with package edgeR. Original data from
  gene expression in the plant species \emph{Arabidopsis thaliana}.
}
\examples{
colnames(volcano_example.df)
head(volcano_example.df)

}
\references{
Rai, Neha; O'Hara, Andrew; Farkas, Daniel; Safronov, Omid;
Ratanasopa, Khuanpiroon; Wang, Fang; Lindfors, Anders V.; Jenkins,
Gareth I.; Lehto, Tarja; Salojärvi, Jarkko; Brosché, Mikael; Strid. Åke;
 Aphalo, Pedro José; Morales, Luis Orlando (2020) The photoreceptor UVR8
 mediates the perception of both UV-B and UV-A wavelengths up to 350 nm of
 sunlight with responsivity moderated by cryptochromes. \emph{Plant, Cell &
 Environment}, 43:1513-1527.
}
\seealso{
Other Transcriptomics data: 
\code{\link{quadrant_example.df}}
}
\concept{Transcriptomics data}
\keyword{datasets}
